Various fixes and improvements to the Lua bindings
- Add bindings for the audio device class
- Add a method to toggle pause in the Lua application class
- Add a method in the Lua drawable node class to get its size
- Add support for loading the engine library with the debug postfix
- Fix the creation methods of the Lua rectangle class
- Fix the exported Lua values of the joystick dead zones
- Fix the export of texture and textnode constants
- Rename a method of the Lua vector classes
- Remove CMake commands to copy the script directory to the data one

+++ src/scripting/LuaIInputManager.cpp
@@ -97,9 +97,9 @@ void LuaIInputManager::exposeConstants(lua_State *L)
 
 	lua_createtable(L, 3, 0);
 
-	LuaUtils::pushField(L, LuaNames::IInputManager::LEFT_STICK, static_cast<int64_t>(IInputManager::LeftStickDeadZone));
-	LuaUtils::pushField(L, LuaNames::IInputManager::RIGHT_STICK, static_cast<int64_t>(IInputManager::RightStickDeadZone));
-	LuaUtils::pushField(L, LuaNames::IInputManager::TRIGGER, static_cast<int64_t>(IInputManager::TriggerDeadZone));
+	LuaUtils::pushField(L, LuaNames::IInputManager::LEFT_STICK, IInputManager::LeftStickDeadZone);
+	LuaUtils::pushField(L, LuaNames::IInputManager::RIGHT_STICK, IInputManager::RightStickDeadZone);
+	LuaUtils::pushField(L, LuaNames::IInputManager::TRIGGER, IInputManager::TriggerDeadZone);
 
 	lua_setfield(L, -2, LuaNames::IInputManager::DeadZone);
 }
